Split status diagnostics tests by family
