+++ src/supervisor/supervisor-status-rendering-supervisor-review-bot.test.ts
@@ -16,6 +16,7 @@ import {
   SupervisorStateFile,
 } from "../core/types";
 
+
 function createConfig(overrides: Partial<SupervisorConfig> = {}): SupervisorConfig {
   return {
     repoPath: "/tmp/repo",
@@ -100,91 +101,6 @@ Parallelizable: No
   };
 }
 
-test("status readiness shows stale-review root blockers through dependency chains", async () => {
-  const config = createConfig();
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      "1695": createRecord({
-        issue_number: 1695,
-        state: "blocked",
-        blocked_reason: "stale_review_bot",
-      }),
-    },
-  };
-  const staleRoot = createExecutionReadyIssue({
-    number: 1695,
-    title: "Refresh configured-bot metadata",
-  });
-  const firstDependent = createExecutionReadyIssue({
-    number: 1696,
-    title: "Use fresh configured-bot metadata",
-    body: `## Summary
-Run only after the configured-bot metadata blocker clears.
-
-## Scope
-- depend on the stale review-bot predecessor
-
-## Acceptance criteria
-- status shows the dependency root blocker
-
-## Verification
-- npx tsx --test src/supervisor/supervisor-status-rendering-supervisor.test.ts
-
-Depends on: #1695
-Parallelizable: No
-
-## Execution order
-1 of 1`,
-  });
-  const secondDependent = createExecutionReadyIssue({
-    number: 1697,
-    title: "Run after the dependent issue",
-    body: `## Summary
-Run only after the previous dependent issue clears.
-
-## Scope
-- depend on the chained predecessor
-
-## Acceptance criteria
-- status shows the dependency root blocker through the chain
-
-## Verification
-- npx tsx --test src/supervisor/supervisor-status-rendering-supervisor.test.ts
-
-Depends on: #1696
-Parallelizable: No
-
-## Execution order
-1 of 1`,
-  });
-
-  const summary = await buildReadinessSummary(
-    {
-      listCandidateIssues: async () => [firstDependent, secondDependent],
-      listAllIssues: async () => [staleRoot, firstDependent, secondDependent],
-    },
-    config,
-    state,
-  );
-
-  assert.deepEqual(summary.blockedIssues, [
-    {
-      issueNumber: 1696,
-      title: "Use fresh configured-bot metadata",
-      blockedBy: "depends on #1695 root_blocker=#1695 blocked_reason=stale_review_bot",
-    },
-    {
-      issueNumber: 1697,
-      title: "Run after the dependent issue",
-      blockedBy: "depends on #1696 root_blocker=#1695 blocked_reason=stale_review_bot",
-    },
-  ]);
-  assert.match(
-    summary.readinessLines.join("\n"),
-    /blocked_issues=#1696 blocked_by=depends on #1695 root_blocker=#1695 blocked_reason=stale_review_bot; #1697 blocked_by=depends on #1696 root_blocker=#1695 blocked_reason=stale_review_bot/,
-  );
-});
 
 function createRecord(overrides: Partial<IssueRunRecord> = {}): IssueRunRecord {
   return {
@@ -299,108 +215,6 @@ function withStubbedDateNow<T>(nowIso: string, run: () => T): T {
   }
 }
 
-test("summarizeChecks treats cancelled runs as waiting, not failing", () => {
-  const summary = summarizeChecks([{ name: "merge-queue", state: "CANCELLED", bucket: "cancel", workflow: "CI" }]);
-
-  assert.equal(summary.allPassing, false);
-  assert.equal(summary.hasPending, true);
-  assert.equal(summary.hasFailing, false);
-});
-
-test("formatDetailedStatus preserves conversation-resolution blocker status line tokens", () => {
-  const status = formatDetailedStatus({
-    config: createConfig({
-      reviewBotLogins: ["chatgpt-codex-connector"],
-    }),
-    activeRecord: createRecord({
-      pr_number: 42,
-      state: "pr_open",
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr: createPullRequest({
-      mergeStateStatus: "BLOCKED",
-      mergeable: "MERGEABLE",
-      configuredBotCurrentHeadObservedAt: "2026-05-26T00:00:00Z",
-      configuredBotCurrentHeadStatusState: "SUCCESS",
-      requiredConversationResolution: {
-        state: "enabled",
-        source: "branch_protection",
-        details: ["required_conversation_resolution=enabled"],
-      },
-    }),
-    checks: [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }],
-    reviewThreads: [
-      createReviewThread({
-        id: "thread-2",
-        isOutdated: true,
-      }),
-      createReviewThread({
-        id: "thread-1",
-        isOutdated: true,
-      }),
-    ],
-  });
-
-  assert.match(
-    status,
-    /^conversation_resolution_blocker state=blocked required_conversation_resolution=enabled outdated_configured_bot_threads=2 thread_ids=thread-1,thread-2$/m,
-  );
-});
-
-test("formatDetailedStatus shows blocking local review status for current PR head", () => {
-  const config = createConfig({ localReviewEnabled: true, localReviewPolicy: "block_ready" });
-  const record = createRecord({
-    local_review_head_sha: "deadbeef",
-    local_review_blocker_summary: "high src/supervisor.ts:210-214 stale artifact context drives the wrong repair path.",
-    local_review_max_severity: "high",
-    local_review_findings_count: 3,
-    local_review_recommendation: "changes_requested",
-    local_review_run_at: "2026-03-11T14:05:00Z",
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr: createPullRequest({ isDraft: true }),
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /local_review gating=yes policy=block_ready findings=3 root_causes=0 max_severity=high verified_findings=0 verified_max_severity=none head=current reviewed_head_sha=deadbeef pr_head_sha=deadbeef ran_at=2026-03-11T14:05:00Z blocker_summary=high src\/supervisor\.ts:210-214 stale artifact context drives the wrong repair path\. signature=none repeated=0 stalled=no/,
-  );
-  assert.doesNotMatch(status, /needs_review_run=/);
-  assert.match(status, /external_review head=none reviewed_head_sha=none matched=0 near_match=0 missed=0/);
-});
-
-test("formatDetailedStatus shows both raw and compressed local review counts", () => {
-  const config = createConfig({ localReviewPolicy: "block_ready" });
-  const record = createRecord({
-    local_review_head_sha: "deadbeef",
-    local_review_max_severity: "high",
-    local_review_findings_count: 3,
-    local_review_root_cause_count: 1,
-    local_review_recommendation: "changes_requested",
-    local_review_run_at: "2026-03-11T14:05:00Z",
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr: createPullRequest({ isDraft: true }),
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /local_review .*findings=3 .*root_causes=1 .*stalled=no/);
-});
-
 test("formatDetailedStatus shows configured-bot same-head follow-up eligibility and exhaustion", () => {
   const config = createConfig({ reviewBotLogins: ["copilot-pull-request-reviewer"] });
   const pr = createPullRequest({
@@ -462,121 +276,6 @@ test("formatDetailedStatus shows configured-bot same-head follow-up eligibility 
     reviewThreads,
   });
   assert.match(exhaustedStatus, /review_follow_up state=exhausted remaining=0 head_sha=deadbeef actionable=0/);
-});
-
-test("buildDetailedStatusSummaryLines surfaces same-PR follow-up repair without changing the saved outcome", () => {
-  const statusLines = buildDetailedStatusSummaryLines({
-    config: createConfig({ localReviewFollowUpRepairEnabled: true }),
-    activeRecord: createRecord({
-      state: "local_review_fix",
-      local_review_summary_path: "/tmp/reviews/owner-repo/issue-366/head-deadbeef.md",
-    }),
-    activityContext: {
-      handoffSummary: null,
-      localReviewRoutingSummary: null,
-      changeClassesSummary: null,
-      verificationPolicySummary: null,
-      durableGuardrailSummary: null,
-      externalReviewFollowUpSummary: null,
-      preMergeEvaluation: {
-        status: "follow_up_eligible",
-        outcome: "follow_up_eligible",
-        repair: "same_pr_follow_up_current_head",
-        reason: "follow_up_candidates=1",
-        headStatus: "current",
-        summaryPath: "owner-repo/issue-366/head-deadbeef.md",
-        artifactPath: "owner-repo/issue-366/head-deadbeef.json",
-        ranAt: "2026-03-24T00:11:00Z",
-        mustFixCount: 0,
-        manualReviewCount: 0,
-        followUpCount: 1,
-      },
-      localCiStatus: null,
-      latestRecovery: null,
-      retryContext: {
-        timeoutRetryCount: 0,
-        blockedVerificationRetryCount: 0,
-        repeatedBlockerCount: 0,
-        repeatedFailureSignatureCount: 0,
-        lastFailureSignature: null,
-      },
-      repeatedRecovery: null,
-      recentPhaseChanges: [],
-      localReviewSummaryPath: "owner-repo/issue-366/head-deadbeef.md",
-      externalReviewMissesPath: null,
-      reviewWaits: [],
-    },
-  });
-
-  assert.ok(
-    statusLines.includes(
-      "pre_merge_evaluation status=follow_up_eligible outcome=follow_up_eligible repair=same_pr_follow_up_current_head head=current must_fix=0 manual_review=0 follow_up=1 reason=follow_up_candidates=1 ran_at=2026-03-24T00:11:00Z summary_path=owner-repo/issue-366/head-deadbeef.md artifact_path=owner-repo/issue-366/head-deadbeef.json",
-    ),
-  );
-});
-
-test("formatDetailedStatus marks stalled local-review repair loops explicitly", () => {
-  const config = createConfig({
-    localReviewEnabled: true,
-    localReviewPolicy: "block_ready",
-    localReviewHighSeverityAction: "retry",
-    sameFailureSignatureRepeatLimit: 3,
-  });
-  const record = createRecord({
-    state: "blocked",
-    blocked_reason: "verification",
-    local_review_head_sha: "deadbeef",
-    local_review_max_severity: "high",
-    local_review_verified_max_severity: "high",
-    local_review_verified_findings_count: 1,
-    local_review_findings_count: 3,
-    local_review_root_cause_count: 1,
-    local_review_recommendation: "changes_requested",
-    repeated_local_review_signature_count: 3,
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr: createPullRequest({ isDraft: true }),
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /local_review .* repeated=3 stalled=yes/);
-  assert.match(status, /blocked_reason=verification/);
-});
-
-test("formatDetailedStatus shows saved external review miss counts for the current PR head", () => {
-  const config = createConfig();
-  const pr = createPullRequest({
-    number: 22,
-    title: "Add review learning",
-    headRefName: "codex/issue-58",
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: createRecord({
-      pr_number: 22,
-      state: "addressing_review",
-      external_review_head_sha: "deadbeef",
-      external_review_misses_path: "/tmp/reviews/owner-repo/issue-58/external-review-misses-head-deadbeef.json",
-      external_review_matched_findings_count: 1,
-      external_review_near_match_findings_count: 1,
-      external_review_missed_findings_count: 2,
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /external_review head=current reviewed_head_sha=deadbeef matched=1 near_match=1 missed=2/);
-  assert.match(status, /external_review_misses_path=owner-repo\/issue-58\/external-review-misses-head-deadbeef\.json/);
 });
 
 test("formatDetailedStatus surfaces not_requested Copilot review lifecycle", () => {
@@ -737,37 +436,6 @@ test("formatDetailedStatus surfaces configured bot review timeout outcome with g
     status,
     /timeout_reason=Requested configured review bot \(chatgpt-codex-connector\) review never arrived within 10 minute\(s\) for head deadbeef\./,
   );
-});
-
-test("formatDetailedStatus surfaces final auto-merge guard evidence", () => {
-  const status = formatDetailedStatus({
-    config: createConfig(),
-    activeRecord: createRecord({
-      pr_number: 44,
-      state: "merging",
-      last_auto_merge_guard_context: {
-        category: null,
-        summary: "Final auto-merge guard passed for PR #44.",
-        signature: "auto-merge-ready:head-44",
-        command: null,
-        details: ["head_sha=head-44", "checks=green count=1", "configured_bot_blockers=0"],
-        url: "https://example.test/pr/44",
-        updated_at: "2026-03-13T06:30:00Z",
-      },
-    }),
-    latestRecord: null,
-    trackedIssueCount: 1,
-    pr: createPullRequest({
-      number: 44,
-      headRefName: "codex/issue-38",
-      headRefOid: "head-44",
-    }),
-    checks: [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /auto_merge_guard summary=Final auto-merge guard passed for PR #44\./);
-  assert.match(status, /auto_merge_guard_details=head_sha=head-44 \| checks=green count=1 \| configured_bot_blockers=0/);
 });
 
 test("formatDetailedStatus explains softened nitpick-only configured-bot top-level reviews", () => {
@@ -1163,167 +831,4 @@ test("formatDetailedStatus keeps generic configured bot timeout wording for mixe
     status,
     /timeout_reason=Requested configured review bots \(copilot-pull-request-reviewer, chatgpt-codex-connector\) review never arrived within 10 minute\(s\) for head deadbeef\./,
   );
-});
-
-test("formatDetailedStatus surfaces the latest recovery reason separately from the active issue", () => {
-  const config = createConfig();
-  const activeRecord = createRecord({
-    issue_number: 92,
-    state: "implementing",
-    branch: "codex/issue-92",
-    workspace: "/tmp/workspaces/issue-92",
-    blocked_reason: null,
-    last_error: null,
-    last_failure_kind: null,
-    last_failure_context: null,
-    last_failure_signature: null,
-    codex_session_id: null,
-  });
-  const latestRecoveryRecord = createRecord({
-    issue_number: 91,
-    state: "done",
-    branch: "codex/issue-91",
-    workspace: "/tmp/workspaces/issue-91",
-    blocked_reason: null,
-    last_error: null,
-    last_failure_kind: null,
-    last_failure_context: null,
-    last_failure_signature: null,
-    codex_session_id: null,
-    updated_at: "2026-03-13T00:20:00Z",
-    last_codex_summary: null,
-    last_recovery_reason: "merged_pr_convergence: tracked PR #191 merged; marked issue #91 done",
-    last_recovery_at: "2026-03-13T00:20:00Z",
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord,
-    latestRecord: latestRecoveryRecord,
-    latestRecoveryRecord,
-    trackedIssueCount: 2,
-    pr: null,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /latest_recovery issue=#91 at=2026-03-13T00:20:00Z reason=merged_pr_convergence detail=tracked PR #191 merged; marked issue #91 done/,
-  );
-});
-
-test("formatDetailedStatus reports idle status with the latest record and latest recovery", () => {
-  const config = createConfig();
-  const latestRecord = createRecord({
-    issue_number: 92,
-    state: "done",
-    branch: "codex/issue-92",
-    updated_at: "2026-03-13T01:20:00Z",
-  });
-  const latestRecoveryRecord = createRecord({
-    issue_number: 91,
-    state: "done",
-    branch: "codex/issue-91",
-    workspace: "/tmp/workspaces/issue-91",
-    updated_at: "2026-03-13T00:20:00Z",
-    last_codex_summary: null,
-    last_recovery_reason: "merged_pr_convergence: tracked PR #191 merged; marked issue #91 done",
-    last_recovery_at: "2026-03-13T00:20:00Z",
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: null,
-    latestRecord,
-    latestRecoveryRecord,
-    trackedIssueCount: 2,
-    pr: null,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(status, /^No active issue\./);
-  assert.match(status, /tracked_issues=2/);
-  assert.match(status, /latest_record=#92 state=done updated_at=2026-03-13T01:20:00Z/);
-  assert.match(
-    status,
-    /latest_recovery issue=#91 at=2026-03-13T00:20:00Z reason=merged_pr_convergence detail=tracked PR #191 merged; marked issue #91 done/,
-  );
-});
-
-test("formatDetailedStatus marks stale local review as gating until current-head final evaluation resolves", () => {
-  const config = createConfig({ localReviewEnabled: true, localReviewPolicy: "block_merge" });
-  const record = createRecord({
-    local_review_head_sha: "oldhead",
-    local_review_max_severity: "medium",
-    local_review_findings_count: 2,
-    local_review_recommendation: "changes_requested",
-    local_review_run_at: "2026-03-11T14:05:00Z",
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr: createPullRequest({ headRefOid: "newhead" }),
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /local_review gating=yes policy=block_merge findings=2 root_causes=0 max_severity=medium verified_findings=0 verified_max_severity=none head=stale reviewed_head_sha=oldhead pr_head_sha=newhead ran_at=2026-03-11T14:05:00Z needs_review_run=yes drift=oldhead->newhead/,
-  );
-});
-
-test("formatDetailedStatus reports unknown local review head status without a PR", () => {
-  const config = createConfig({ localReviewPolicy: "block_merge" });
-  const record = createRecord({
-    local_review_head_sha: "oldhead",
-    local_review_max_severity: "medium",
-    local_review_findings_count: 2,
-    local_review_recommendation: "changes_requested",
-    local_review_run_at: "2026-03-11T14:05:00Z",
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr: null,
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /local_review gating=no policy=block_merge findings=2 root_causes=0 max_severity=medium verified_findings=0 verified_max_severity=none head=unknown reviewed_head_sha=oldhead pr_head_sha=unknown ran_at=2026-03-11T14:05:00Z/,
-  );
-});
-
-test("formatDetailedStatus reports none local review head status with current PR head", () => {
-  const config = createConfig({ localReviewEnabled: true, localReviewPolicy: "block_merge" });
-  const record = createRecord({
-    local_review_head_sha: null,
-    local_review_run_at: null,
-  });
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr: createPullRequest({ headRefOid: "newhead" }),
-    checks: [],
-    reviewThreads: [],
-  });
-
-  assert.match(
-    status,
-    /local_review gating=yes policy=block_merge findings=0 root_causes=0 max_severity=none verified_findings=0 verified_max_severity=none head=none reviewed_head_sha=none pr_head_sha=newhead ran_at=none/,
-  );
-  assert.doesNotMatch(status, /blocker_summary=/);
 });
